Format README file
I thought something was going to have to change in this file, so I went
through it adding formatting, looking for the section that would change.
I guess this wasn't necessary since I didn't actually change any text in
the instructions. But I think this'll be more readable for new users, so
committing it anyway.