Added Comment on show/hide script
I forgot to comment that the the show/hide counter script really does
need to be located in a file, not embedded in the page body. I thought
the script.html file might be the answer, but I couldn’t get the code
to show up in the head section.