Fixed problem of no anchor tags in leader board
The documentation of the the leader board with ‘whatshot-with-links’
theme suggests this is sufficient to cause the name and avatar to be
hyperlinked to a particular stream, and replaces the first stream with
the hyperlinked stream. In reality, the leader board depends on a tab
module being present (not a simple stream). One tab is sufficient. The
leader board links affect the stream used in the tab module.