Changed appearance of leader-board
Changed the column title from <strong> to <h4> to get the advantage of
the CSS. I noticed some weird behavior if not all of the celebrities
were shown and the ‘more’ was clicked; sometimes one celebrity would
show up twice, or a bar would be filled disproportionately to the
others. Showing all of the celebrities (popular female vocalists)
eliminates this problem. Should still be mentioned to Arktan. Lastly,
added comment about the necessity of having a ‘tab’ app for a clickable
link in the leader-board, and how the trend module must follow the tabs
module.