Initial push of Ratings and Review
Created the page partially from Arktan sandbox example and from Eric’s
Ratings and Reviews page on Appspot. This was working partially, but
the main problem appears to be in the login and logout section. Login
was working, but logout never did. Pushing up for anybody to take a
look at.