Changed the ppm counter filter to 'all'.
I reduced the stream to just 6 twitter searches, and the set the trend
to watch that stream. The numbers in the counter and the final numbers
in the charge are still off by a factor of 2-3. Mostly I just want to
make sure the latest PPM page is in place.