Switch counters to 'percentage' from 'number'
I forgot that the values appearing in the little counter areas on
either side of the hashtag battle ‘speedometer’ display consistently if
they are ‘percentage’ values instead of raw counts. This is because if
the count size is too large, the display of the numbers may be hidden
or wrap.